+++ src/lib.rs
@@ -32,8 +32,8 @@
 //!
 //! let bus = shared_bus::BusManagerSimple::new(i2c);
 //!
-//! let mut proxy1 = bus.acquire_i2c();
-//! let mut my_device = MyDevice::new(bus.acquire_i2c());
+//! let mut proxy1 = bus.acquire();
+//! let mut my_device = MyDevice::new(bus.acquire());
 //!
 //! proxy1.write(0x39, &[0xc0, 0xff, 0xee]);
 //! my_device.do_something_on_the_bus();
@@ -67,8 +67,8 @@
 //! // shared with other threads.
 //! let bus: &'static _ = shared_bus::new_std!(SomeI2cBus = i2c).unwrap();
 //!
-//! let mut proxy1 = bus.acquire_i2c();
-//! let mut my_device = MyDevice::new(bus.acquire_i2c());
+//! let mut proxy1 = bus.acquire();
+//! let mut my_device = MyDevice::new(bus.acquire());
 //!
 //! // We can easily move a proxy to another thread:
 //! # let t =
@@ -93,23 +93,21 @@
 //!
 //! | Bus/Block | Proxy Type | Acquire Method | Comments |
 //! | --- | --- | --- | --- |
-//! | I2C | [`I2cProxy`] | [`.acquire_i2c()`] | |
+//! | I2C | [`Proxy`] | [`.acquire()`] | |
+//! | ADC | [`Proxy`] | [`.acquire()`] | |
 //! | SPI | [`SpiProxy`] | [`.acquire_spi()`] | SPI can only be shared within a single task (See [`SpiProxy`] for details). |
-//! | ADC | [`AdcProxy`] | [`.acquire_adc()`] | |
 //!
 //!
-//! [`.acquire_i2c()`]: ./struct.BusManager.html#method.acquire_i2c
+//! [`.acquire()`]: ./struct.BusManager.html#method.acquire
 //! [`.acquire_spi()`]: ./struct.BusManager.html#method.acquire_spi
-//! [`.acquire_adc()`]: ./struct.BusManager.html#method.acquire_adc
 //! [`BusManagerCortexM`]: ./type.BusManagerCortexM.html
 //! [`BusManagerXtensa`]: ./type.BusManagerXtensa.html
 //! [`BusManagerAtomicCheck`]: ./type.BusManagerAtomicCheck.html
 //! [`BusManagerSimple`]: ./type.BusManagerSimple.html
 //! [`BusManagerStd`]: ./type.BusManagerStd.html
 //! [`BusMutex`]: ./trait.BusMutex.html
-//! [`I2cProxy`]: ./struct.I2cProxy.html
+//! [`Proxy`]: ./struct.Proxy.html
 //! [`SpiProxy`]: ./struct.SpiProxy.html
-//! [`AdcProxy`]: ./struct.AdcProxy.html
 //! [`new_cortexm!()`]: ./macro.new_cortexm.html
 //! [`new_xtensa!()`]: ./macro.new_xtensa.html
 //! [`new_std!()`]: ./macro.new_std.html
@@ -143,8 +141,11 @@ pub use mutex::CortexMMutex;
 pub use mutex::NullMutex;
 #[cfg(feature = "xtensa")]
 pub use mutex::XtensaMutex;
-pub use proxies::AdcProxy;
+pub use proxies::Proxy;
+#[allow(deprecated)]
 pub use proxies::I2cProxy;
+#[allow(deprecated)]
+pub use proxies::AdcProxy;
 pub use proxies::SpiProxy;
 
 #[cfg(feature = "cortex-m")]
@@ -174,8 +175,8 @@ pub use mutex::AtomicCheckMutex;
 ///
 /// let bus = shared_bus::BusManagerSimple::new(i2c);
 ///
-/// let mut proxy1 = bus.acquire_i2c();
-/// let mut my_device = MyDevice::new(bus.acquire_i2c());
+/// let mut proxy1 = bus.acquire();
+/// let mut my_device = MyDevice::new(bus.acquire());
 ///
 /// proxy1.write(0x39, &[0xc0, 0xff, 0xee]);
 /// my_device.do_something_on_the_bus();
@@ -248,7 +249,7 @@ pub type BusManagerXtensa<BUS> = BusManager<XtensaMutex<BUS>>;
 ///
 /// // the HAL I2C driver type
 /// type I2cType = ();
-/// type Proxy = shared_bus::I2cProxy<'static, shared_bus::AtomicCheckMutex<I2cType>>;
+/// type Proxy = shared_bus::Proxy<'static, shared_bus::AtomicCheckMutex<I2cType>>;
 ///
 /// struct SharedBusDevices {
 ///     device: Device<Proxy>,
@@ -268,8 +269,8 @@ pub type BusManagerXtensa<BUS> = BusManager<XtensaMutex<BUS>>;
 ///     let bus_manager: &'static _ = shared_bus::new_atomic_check!(I2cType = i2c).unwrap();
 ///
 ///     let devices = SharedBusDevices {
-///         device: Device { bus: bus_manager.acquire_i2c() },
-///         other_device: OtherDevice { bus: bus_manager.acquire_i2c() },
+///         device: Device { bus: bus_manager.acquire() },
+///         other_device: OtherDevice { bus: bus_manager.acquire() },
 ///     };
 ///
 ///     Resources {